+++ Twia.AzureFunctions.Extensions.OpenApi.UnitTests/HttpFunctionPathParameterTypeFilterTests.cs
@@ -64,7 +64,7 @@ namespace Twia.AzureFunctions.Extensions.OpenApi.UnitTests
                         In = param.Location
                     }))
             };
-            var method = typeof(FunctionMethodTestSource).GetMethod(nameof(FunctionMethodTestSource.WithParameters));
+            var method = FunctionMethodTestSource.GetMethodInfo();
             _schemaRegistry = A.Fake<ISchemaGenerator>();
             _schemaRepository = new SchemaRepository();
             var apiDescription = new ApiDescription();
@@ -132,14 +132,6 @@ namespace Twia.AzureFunctions.Extensions.OpenApi.UnitTests
                 {
                     openApiParameter.Schema.Should().BeNull("Schema should not be set for non-Path parameters.");
                 }
-            }
-        }
-
-        private static class FunctionMethodTestSource
-        {
-            public static void WithParameters(string stringParam, int? integerParam, bool booleanParam)
-            {
-                // Just for testing.
             }
         }
     }
